+++ libs/extract/src/lib/dropdown/dropdown.ts
@@ -11,6 +11,8 @@ import {
   popper,
   keypress,
   validate,
+  search,
+  selectByValue,
 } from './reducers'
 import { fromEvent, merge, Subject } from 'rxjs'
 import { take, takeUntil } from 'rxjs/operators'
@@ -19,27 +21,30 @@ import {
   DropdownHandler,
   DropdownListener,
   DropdownArgs,
+  OnChange,
 } from './types'
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock'
 import { isMobileViewport$ } from '../common/viewport-size'
 import { getOptionScrollPosition } from './helper-functions'
+import { IValidator } from '../helperFunction'
 
 export const createDropdown = (
   init: DropdownArgs,
   toggler: HTMLElement,
   listbox: HTMLElement,
   fieldset: HTMLElement | undefined = undefined,
-  listener: DropdownListener
+  listener: DropdownListener,
+  onChange: OnChange
 ): DropdownHandler => {
-  const _handler: Partial<DropdownHandler> = {
+  const handler: DropdownHandler = {
     toggler,
     listbox,
     dropdown: create(init),
     isAlive: true,
+    onChange,
     onDestroy$: new Subject<void>(),
     onTouched: init.onTouched,
-  }
-  const handler = _handler as DropdownHandler
+  } as DropdownHandler
 
   handler.active = (isActive) =>
     update(handler, listener, active(handler.dropdown, isActive))
@@ -49,6 +54,8 @@ export const createDropdown = (
   handler.open = () => update(handler, listener, open(handler.dropdown))
   handler.close = () => update(handler, listener, close(handler.dropdown))
   handler.toggle = () => update(handler, listener, toggle(handler.dropdown))
+  handler.search = (seachInput: string) =>
+    update(handler, listener, search(handler.dropdown, seachInput))
   handler.select = (selection, selectOnClose = true) =>
     update(
       handler,
@@ -57,8 +64,10 @@ export const createDropdown = (
         ? close(select(handler.dropdown, selection))
         : select(handler.dropdown, selection)
     )
+  handler.selectByValue = (selection) =>
+    update(handler, listener, selectByValue(handler.dropdown, selection))
   handler.update = (props) => update(handler, listener, create(props))
-  handler.validate = (validator: any) =>
+  handler.validate = (validator: IValidator) =>
     update(handler, listener, validate(handler.dropdown, validator))
 
   fromEvent(toggler, 'blur')
@@ -141,6 +150,12 @@ const update = async (
 
   const oldState = handler.dropdown
   if (newState) handler.dropdown = newState
+
+  if (
+    JSON.stringify(oldState.value) !== JSON.stringify(handler.dropdown.value)
+  ) {
+    handler.onChange(handler.dropdown.value)
+  }
 
   if (oldState.isTouched !== handler.dropdown.isTouched) {
     handler.onTouched?.()