+++ libs/extract/src/lib/dropdown/types.ts
@@ -1,12 +1,20 @@
 import { Subscription, Subject } from 'rxjs'
 import { AriaAttributes } from 'react'
 import { Instance } from '@popperjs/core'
+import { IValidator } from '../helperFunction'
+
+export type CompareWith<T = any> = (o1: T, o2: T) => boolean
+
+export type SearchFilter<T = any> = (search: string, value: T) => boolean
+
+export type OnChange<T = any> = (value: T) => void
+
+export type DropdownListener = (dropdown: AbstractDropdown) => void
 
 export interface DropdownOption {
-  key?: string
-  value?: unknown
+  label?: string
+  value?: any
   selected?: boolean
-  active?: boolean
   [key: string]: any
 }
 
@@ -14,11 +22,15 @@ export interface DropdownTexts {
   select?: string
   selected?: string
   placeholder?: string
+  searchPlaceholder?: string
   close?: string
   optionsDescription?: string
 }
 
-export interface ExtendedDropdownOption extends DropdownOption, ElementProps {}
+export interface DropdownOptionElement extends DropdownOption, ElementProps {
+  active?: boolean
+  hidden?: boolean
+}
 
 export interface Attributes extends AriaAttributes {
   id?: string
@@ -34,17 +46,20 @@ export interface ElementProps {
 
 export interface AbstractDropdown {
   id: string
+  value: any
   texts: DropdownTexts
   isActive: boolean
   isOpen: boolean
   isLooping: boolean
   isMultiSelect: boolean
+  isSearchable: boolean
   isTouched: boolean
   useValue: string
   display: string
-  selectValue: string
-  validator: any
-  options: ExtendedDropdownOption[]
+  validator: IValidator
+  compareWith: CompareWith
+  searchFilter?: SearchFilter
+  options: DropdownOptionElement[]
   elements: Partial<{
     toggler: Partial<ElementProps>
     listbox: Partial<ElementProps>
@@ -52,20 +67,20 @@ export interface AbstractDropdown {
   }>
 }
 
-export type DropdownListener = (dropdown: AbstractDropdown) => void
-
 export interface DropdownArgs {
   id?: string
   options: DropdownOption[]
   useValue?: string // option key to use as value
   display?: string // option key to display
-  selectValue?: string // option key to output as value
   loop?: boolean
   texts?: DropdownTexts
   value?: any
   multiSelect?: boolean
+  searchable?: boolean
   onTouched?: () => void
-  validator?: any
+  validator?: IValidator
+  compareWith?: CompareWith
+  searchFilter?: SearchFilter
 }
 
 export interface DropdownHandler {
@@ -75,6 +90,7 @@ export interface DropdownHandler {
   popper?: Instance
   subscription: Subscription
   isAlive: boolean
+  onChange: OnChange
   onDestroy$: Subject<void>
   onTouched?: () => void
 
@@ -83,13 +99,16 @@ export interface DropdownHandler {
   active: (isActive: boolean) => Promise<void>
   loop: (isLooping: boolean) => Promise<void>
   multiSelect: (isMultiSelect: boolean) => Promise<void>
+  searchable: (isSearchable: boolean) => Promise<void>
   open: () => Promise<void>
   close: () => Promise<void>
   toggle: () => Promise<void>
   select: (
-    selection: ExtendedDropdownOption,
+    selection: DropdownOptionElement,
     closeOnSelect?: boolean
   ) => Promise<void>
-  validate: (validator: any) => Promise<void>
+  selectByValue: (selection: any) => Promise<void>
+  validate: (validator: IValidator) => Promise<void>
+  search: (searchInput: string) => Promise<void>
   destroy: () => void
 }