+++ libs/charts/CHANGELOG.md
@@ -2,6 +2,70 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.0.0-beta.33](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.32...@sebgroup/green-charts@1.0.0-beta.33) (2022-11-11)
+
+
+
+# [1.0.0-beta.32](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.31...@sebgroup/green-charts@1.0.0-beta.32) (2022-11-10)
+
+
+
+# [1.0.0-beta.31](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.30...@sebgroup/green-charts@1.0.0-beta.31) (2022-11-08)
+
+
+
+# [1.0.0-beta.30](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.29...@sebgroup/green-charts@1.0.0-beta.30) (2022-11-08)
+
+
+
+# [1.0.0-beta.29](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.28...@sebgroup/green-charts@1.0.0-beta.29) (2022-11-08)
+
+
+
+# [1.0.0-beta.28](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.27...@sebgroup/green-charts@1.0.0-beta.28) (2022-10-26)
+
+
+
+# [1.0.0-beta.27](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.26...@sebgroup/green-charts@1.0.0-beta.27) (2022-10-12)
+
+
+
+# [1.0.0-beta.26](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.25...@sebgroup/green-charts@1.0.0-beta.26) (2022-10-07)
+
+
+### Features
+
+* allow custom tick values in charts ([a7c615f](https://github.com/sebgroup/green/commit/a7c615fe2194eec114e6e8176b5df658a12cc7e2)), closes [#601](https://github.com/sebgroup/green/issues/601)
+* allow formatting tick labels ([90d4e9f](https://github.com/sebgroup/green/commit/90d4e9f24ae32be4b6c48514aad13d5aeaea7e08)), closes [#602](https://github.com/sebgroup/green/issues/602)
+
+
+
+# [1.0.0-beta.25](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.24...@sebgroup/green-charts@1.0.0-beta.25) (2022-10-03)
+
+
+### Features
+
+* **charts:** add number formatting to tooltip ([3c4c7f2](https://github.com/sebgroup/green/commit/3c4c7f26685e5d08827ec3d7514c1270b7c82810))
+* **charts:** add ticksCount and stepSize options ([590f738](https://github.com/sebgroup/green/commit/590f738d2a652f4ebb09506656756f033cd26894)), closes [#556](https://github.com/sebgroup/green/issues/556)
+
+
+
+# [1.0.0-beta.24](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.23...@sebgroup/green-charts@1.0.0-beta.24) (2022-09-27)
+
+
+
+# [1.0.0-beta.23](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.22...@sebgroup/green-charts@1.0.0-beta.23) (2022-09-23)
+
+
+
+# [1.0.0-beta.22](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.21...@sebgroup/green-charts@1.0.0-beta.22) (2022-09-20)
+
+
+
+# [1.0.0-beta.21](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.20...@sebgroup/green-charts@1.0.0-beta.21) (2022-09-20)
+
+
+
 # [1.0.0-beta.20](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.19...@sebgroup/green-charts@1.0.0-beta.20) (2022-08-31)